Extract Graph and ContentGraph into @parcel/graph package
