+++ packages/core/core/src/requests/AssetGraphRequest.js
@@ -1,8 +1,9 @@
 // @flow strict-local
 
+import type {Diagnostic} from '@parcel/diagnostic';
+import type {NodeId} from '@parcel/graph';
 import type {Async, Symbol, Meta} from '@parcel/types';
 import type {SharedReference} from '@parcel/workers';
-import type {Diagnostic} from '@parcel/diagnostic';
 import type {
   Asset,
   AssetGroup,
@@ -12,7 +13,6 @@ import type {
   DependencyNode,
   Entry,
   InternalSourceLocation,
-  NodeId,
   ParcelOptions,
   Target,
 } from '../types';
