+++ packages/core/core/test/PublicBundle.test.js
@@ -2,11 +2,12 @@
 import type {Bundle as InternalBundle} from '../src/types';
 
 import assert from 'assert';
+import {ContentGraph} from '@parcel/graph';
+
 import {Bundle, NamedBundle, PackagedBundle} from '../src/public/Bundle';
 import BundleGraph from '../src/BundleGraph';
 import {createEnvironment} from '../src/Environment';
 import {DEFAULT_OPTIONS} from './test-utils';
-import ContentGraph from '../src/ContentGraph';
 import {toProjectPath} from '../src/projectPath';
 
 describe('Public Bundle', () => {
