+++ packages/core/graph/src/index.js
@@ -1,0 +1,6 @@
+// @flow strict-local
+
+export type {NodeId, ContentKey, Edge} from './types';
+export {toNodeId, fromNodeId} from './types';
+export {default as Graph, ALL_EDGE_TYPES, GraphOpts, mapVisitor} from './Graph';
+export {default as ContentGraph, SerializedContentGraph} from './ContentGraph';
